Fix certificates
